backend/requirements.txt
fastapi
uvicorn[standard]
openai
pydantic
python-multipart
python-dotenv
